ffuuuuu

+++ app/controllers/Application.scala
@@ -26,7 +26,7 @@ import scala.util.Try
 
 case class SubmitData(problem: String, compiler: Int, inline: String)
 
-case class SubmitMessage(id: Int)
+case class SubmitMessage(id: Long)
 
 object SubmitMessage {
   implicit val formatSubmitMessage = Json.format[SubmitMessage]
