+++ app/models/Events.scala
@@ -3,7 +3,7 @@ package models
 import play.api.libs.EventSource.{EventDataExtractor, EventNameExtractor}
 import play.api.libs.json.Json
 
-case class ClarificationRequestState(contest: Int, pending: Int, newRequest: Boolean)
+case class ClarificationRequestState(contest: Long, pending: Int, newRequest: Boolean)
 
 object ClarificationRequestState {
   implicit val format = Json.format[ClarificationRequestState]
